mv terraform_fieldset_to_init_request
